requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
edge-tts
httpx
aiosmtplib
pypdf
Pillow
pytesseract
pytz
ddgs